+++ src/main/java/com/system/libsystem/exceptions/book/BorrowedBookNotFoundException.java
@@ -7,7 +7,7 @@ import lombok.Getter;
 @AllArgsConstructor
 public class BorrowedBookNotFoundException extends RuntimeException {
 
-    private final int id;
+    private final Long id;
 
     @Override
     public String getMessage() {
