+++ src/main/java/com/system/libsystem/helpermodels/UserBook.java
@@ -1,18 +1,13 @@
 package com.system.libsystem.helpermodels;
 
-import com.system.libsystem.entities.author.AuthorEntity;
-import com.system.libsystem.entities.genre.GenreEntity;
 import lombok.Getter;
 import lombok.Setter;
-
-import java.util.HashSet;
-import java.util.Set;
 
 @Getter
 @Setter
 public class UserBook {
     private String title;
-    private int borrowedBookId;
+    private Long borrowedBookId;
     private String bookDetailsLink;
     private String borrowDate;
     private String returnDate;
